Verify a labelled Object

The lbl_assert_labelled() function verifies that the input object is a well-formed labelled vector. lbl_check_labelled() and lbl_test_labelled() serve as analogous functions, adhering to checkmate conventions.

The functions verify the correct class, that var_label(x) is NULL or character, and that the val_labels(x) are either NULL or match the type of x, in line with the (minimal) specification in haven::labelled().

Usage

lbl_check_labelled(x)

lbl_test_labelled(x)

lbl_assert_labelled(x, .var.name = checkmate::vname(x), add = NULL)

Arguments

x

The object to be checked.

.var.name

Name of the checked object to print in assertions. Defaults to the heuristic implemented in checkmate::vname().

add

Collection to store assertion messages. See checkmate::AssertCollection.

Value

A logical value. If the input object is a 'haven_labelled' vector, it returns TRUE; otherwise, it returns an error message indicating the issue with the input.

See also

Examples

# Check if 'fruit_lbl' is a 'haven_labelled' vector
lbl_check_labelled(fruit_lbl)
#> [1] TRUE