Densify a labelled vector

The lbl_densify function takes a labelled vector and assigns new, densely-ranked integer values to each unique value in the vector. The order of the unique values is preserved. This function keeps the labels correctly associated with their respective values throughout the densifying process.

The sub-type (typeof(x)) of the labelled is preserved for double and integer, but character variables (<labelled<character>>) are not supported as densification is not applicable for character strings.

Usage

lbl_densify(x, unlabelled = c("error", "fill"))

Arguments

x

A labelled vector to be densified.

unlabelled

This parameter determines the handling of unlabelled values in the input vector. It can take two values:

"error": Throw an error if there are unlabelled values in the vector.

"fill": Fill in labels for any unlabelled values using lbl_fill() before densifying.

Value

A labelled vector of the same length as the original vector, x, but with values replaced with their dense ranks. The labels are correctly associated with their respective values and the variable label (if present) is preserved.

Examples

  # Densify the veggies vector
  lbl_densify(veggies)
#> <labelled<double>[50]>: The regular veggies are indexed by the prime numbers 2-11
#>  [1] 3 5 1 4 2 3 4 3 1 5 2 1 4 2 3 4 2 1 1 1 5 5 1 1 2 1 5 3 4 2 5 1 5 1 1 4 3 2
#> [39] 2 1 2 5 4 1 2 1 3 1 4 1
#> 
#> Labels:
#>  value    label
#>      1   Carrot
#>      2   Potato
#>      3   Tomato
#>      4 Cucumber
#>      5 Broccoli

  # Densify exotic_veggies, filling in for unlabelled values
  lbl_densify(exotic_veggies, unlabelled = "fill")
#> <labelled<double>[50]>: The exotic veggies are indexed by the prime numbers 2-41. No observations exist for indexes 7 and 41, but 8 unlabelled observations exist (for indexes 2, 8, 16 and 32), and 1 observations is NA
#>  [1]  1 11  3  1 11 11  4  8  4  2  3  2 15  7  1  2  3  7  2  7  7  4  4  6  1
#> [26]  1  1 13 11  9  6 10  9  7  9  7  7  5  8  6  1  8  1 14  3  4 16 12  5  4
#> 
#> Labels:
#>  value     label
#>      1   Bokchoy
#>      2  Celeriac
#>      3       <4>
#>      4   Chayote
#>      5  Eggplant
#>      6       <8>
#>      7    Endive
#>      8    Fennel
#>      9      <16>
#>     10    Jicama
#>     11  Kohlrabi
#>     12      Okra
#>     13 Radicchio
#>     14  Rutabaga
#>     15      <32>
#>     16      Taro
#>     17  Zucchini