Recode labelled variables using either formula notation or lookup tables

These functions recode labelled variables while taking special care to keep the values and the labels (a.k.a. value labels) of the variable in sync. They assume that the value labels are meaningfully related to the values, so when recoding values, the functions take care to recode the levels of the value labels so that the labels continue to track the underlying values.

For recoding using formula or tilde notation, the underlying recoding takes place using dplyr::case_match(), and for recoding using a lookup table (provided either as a named vector/list or as a data.frame) is done using zmisc::lookup(). Both methods can be used in the same call, in that case the formula notation is applied first, followed by the application of the lookup table.

lbl_recode() does the recode operation on the values but updates the levels of the value labels accordingly. This is appropriate before using the variable for plots or statistical analyses. Recoding multiple values to the same value results in corresponding variable labels being combined. Recoding to NA results in corresponding labels to be dropped since NA should not be labelled in standard haven_labelled variables.

lbl_recode_values() is an alias for lbl_recode(), but makes it explicit that focal target of the recoding is the underlying values in the labelled variable.

lbl_recode_labels() recodes the text description of the labels attached to the labelled variable, leaving the underlying values unchanged. This is appropriate to do before using the variable in functions that display the labels, in particular if the output is intended for a final product or publication.

Usage

lbl_recode(.x, ..., .lookup_table = NULL, .sep = " / ")

lbl_recode_values(.x, ..., .lookup_table = NULL, .sep = " / ")

lbl_recode_labels(.x, ..., .lookup_table = NULL)

Arguments

.x

The labelled vector to recode.

...

A recode specification in tilde (formula) syntax. See dplyr::case_match() for specifics about the format of the parameter.

.lookup_table

A recode specification in the form of a lookup table. See zmisc::lookup() for specifics about the format of the parameter.

.sep

A string for how to separate labels that need to be combined.

Value

The re-coded vector.

Examples

lbl_recode(fruit_lbl, 1:2 ~ 1, 3 ~ 2, 4:5 ~ 3)
#> <labelled<double>[50]>
#>  [1] 2 3 3 2 1 1 1 3 1 1 1 1 3 3 1 3 3 3 1 1 2 3 3 3 3 1 1 2 1 2 1 2 1 1 3 2 3 3
#> [39] 2 1 1 1 3 3 1 3 2 1 3 1
#> 
#> Labels:
#>  value          label
#>      1 Apple / Banana
#>      2           Lime
#>      3 Orange / Peach
lbl_recode(fruit_lbl_chr, "d" ~ "z", "e" ~ NA)
#> <labelled<character>[50]>
#>  [1] c  NA z  c  a  b  a  z  a  a  b  a  NA z  b  NA NA z  b  b  c  z  NA NA z 
#> [26] b  b  c  b  c  a  c  b  a  z  c  z  z  c  a  b  b  z  NA b  NA c  a  NA a 
#> 
#> Labels:
#>  value  label
#>      a  Apple
#>      b Banana
#>      c   Lime
#>      z Orange
lbl_recode(fruit_lbl_chr, .lookup_table = c(a = "b", e = "d"))
#> <labelled<character>[50]>
#>  [1] c d d c b b b d b b b b d d b d d d b b c d d d d b b c b c b c b b d c d d
#> [39] c b b b d d b d c b d b
#> 
#> Labels:
#>  value          label
#>      b Apple / Banana
#>      c           Lime
#>      d Orange / Peach

exotic_recoded <- lbl_recode(exotic_veggies,
                             3:10 ~ 1, 11:20 ~ 2, 21:41 ~ 3,
                             2 ~ NA, NA ~ 99)
lbl_count(exotic_recoded)
#> # A tibble: 5 × 3
#>       v l                                                 n
#>   <dbl> <chr>                                         <dbl>
#> 1     1 Celeriac / Chayote / Eggplant                    16
#> 2     2 Endive / Fennel / Jicama / Kohlrabi              17
#> 3     3 Okra / Radicchio / Rutabaga / Taro / Zucchini     8
#> 4    99 NA                                                1
#> 5    NA NA                                                8

lbl_recode_labels(veggies,
    "Carrot" ~ "A straight underground veggie",
    "Potato" ~ "A round underground veggie",
    "Tomato" ~ "A round overground veggie",
    "Cucumber" ~ "A (mostly) straight overground veggie")
#> <labelled<double>[50]>: The regular veggies are indexed by the prime numbers 2-11
#>  [1]  5 11  2  7  3  5  7  5  2 11  3  2  7  3  5  7  3  2  2  2 11 11  2  2  3
#> [26]  2 11  5  7  3 11  2 11  2  2  7  5  3  3  2  3 11  7  2  3  2  5  2  7  2
#> 
#> Labels:
#>  value                                 label
#>      2         A straight underground veggie
#>      3            A round underground veggie
#>      5             A round overground veggie
#>      7 A (mostly) straight overground veggie
#>     11                              Broccoli

lbl_recode_labels(veggies,
    .lookup_table = c("Tomato" = "Tomaaato"))
#> <labelled<double>[50]>: The regular veggies are indexed by the prime numbers 2-11
#>  [1]  5 11  2  7  3  5  7  5  2 11  3  2  7  3  5  7  3  2  2  2 11 11  2  2  3
#> [26]  2 11  5  7  3 11  2 11  2  2  7  5  3  3  2  3 11  7  2  3  2  5  2  7  2
#> 
#> Labels:
#>  value    label
#>      2   Carrot
#>      3   Potato
#>      5 Tomaaato
#>      7 Cucumber
#>     11 Broccoli