Sort value labels of a labelled variable

Sorts the labels of a labelled variable so that the inner representation of the vector is ascending.

Usage

lbl_sort_labels(x)

Arguments

x

A labelled variable.

Value

A labelled variable with values and label levels equal to x, but with inner representation of the value labels in ascending order.

Examples

lbl_sort_labels(fruit_lbl)
#> <labelled<double>[50]>
#>  [1] 3 5 4 3 1 2 1 4 1 1 2 1 5 4 2 5 5 4 2 2 3 4 5 5 4 2 2 3 2 3 1 3 2 1 4 3 4 4
#> [39] 3 1 2 2 4 5 2 5 3 1 5 1
#> 
#> Labels:
#>  value  label
#>      1  Apple
#>      2 Banana
#>      3   Lime
#>      4 Orange
#>      5  Peach
lbl_sort_labels(veggies)
#> <labelled<double>[50]>: The regular veggies are indexed by the prime numbers 2-11
#>  [1]  5 11  2  7  3  5  7  5  2 11  3  2  7  3  5  7  3  2  2  2 11 11  2  2  3
#> [26]  2 11  5  7  3 11  2 11  2  2  7  5  3  3  2  3 11  7  2  3  2  5  2  7  2
#> 
#> Labels:
#>  value    label
#>      2   Carrot
#>      3   Potato
#>      5   Tomato
#>      7 Cucumber
#>     11 Broccoli
lbl_sort_labels(exotic_veggies)
#> <labelled<double>[50]>: The exotic veggies are indexed by the prime numbers 2-41. No observations exist for indexes 7 and 41, but 8 unlabelled observations exist (for indexes 2, 8, 16 and 32), and 1 observations is NA
#>  [1]  2 23  4  2 23 23  5 16  5  3  4  3 37 13  2  3  4 13  3 13 13  5  5 11  2
#> [26]  2  2 31 23 17 11 19 17 13 17 13 13  8 16 11  2 16  2 32  4  5 NA 29  8  5
#> 
#> Labels:
#>  value     label
#>      2   Bokchoy
#>      3  Celeriac
#>      5   Chayote
#>      7  Eggplant
#>     11    Endive
#>     13    Fennel
#>     17    Jicama
#>     19  Kohlrabi
#>     23      Okra
#>     29 Radicchio
#>     31  Rutabaga
#>     37      Taro
#>     41  Zucchini