Convert labelled variables to desired types

Six parameters define the conversion rules, values passed to these parameters are in the form of tidyselect column specifications. Any columns passed will be converted according to the conversion rules associated with that parameter.

Usage

lbl_to(
  d,
  ...,
  .zapped,
  .numeric,
  .logical,
  .char_values,
  .char_labels,
  .factor,
  .ordered,
  .unfilled = c("error", "fill", "na")
)

Arguments

d

A data.frame with some labelled variables.

...

Reserved for future use.

.zapped

Any columns specified in this parameter are converted by zapping the value labels, using haven::zap_labels(), converting to the underlying data type of the underlying value vector, either integer, double, or character. No type checking is performed, as the underlying type is returned unchanged.

.numeric

Columns specified here are converted to numeric, with an error if the underlying type is not numeric (i.e. integer or double).

.logical

Specifies conversion to logical, with an error if the conversion is not possible (underlying values must be numeric and all values must be 0, 1, or NA).

.char_values

Specifies conversion to a character vector containing the values of the input variable, with an error if the underlying type is not character.

.char_labels

Specifies conversion to a character vector containing the labels of the input variable, with an error if the underlying type is not character (this is done by first converting to a factor using haven::as_factor(), then to character using as.character()).

.factor

Specifies conversion to factor (using haven::as_factor() for the conversion).

.ordered

Specifies conversion to factor (using haven::as_factor() for the conversion).

.unfilled

A switch specifying the behavior if the input is not "filled" (see lbl_fill()). Currently only the "error" setting is implemented, and thus any unfilled labels must be filled before converting.

Details

The absence of a .character parameter, is intentional, in order to eliminate ambiguity character conversion is handled by two parameters, .char_values and .char_labels.

Matching the same column in two parameters is an error, since the result would be ambiguous.

Examples

d <- tibble::tibble(veggies, fruit_lbl, fruit_lbl_int, fruit_lbl_chr)
lbl_to(d, .char_labels = veggies, .zapped = starts_with("fruit"))
#> # A tibble: 50 × 4
#>    veggies  fruit_lbl fruit_lbl_int fruit_lbl_chr
#>    <chr>        <dbl>         <int> <chr>        
#>  1 Tomato           3             3 c            
#>  2 Broccoli         5             5 e            
#>  3 Carrot           4             4 d            
#>  4 Cucumber         3             3 c            
#>  5 Potato           1             1 a            
#>  6 Tomato           2             2 b            
#>  7 Cucumber         1             1 a            
#>  8 Tomato           4             4 d            
#>  9 Carrot           1             1 a            
#> 10 Broccoli         1             1 a            
#> # ℹ 40 more rows