Recode labelled variables using tilde notation

Recodes labelled variables and takes care to recode the levels of the value labels so that the labels continue to track the underlying values. Recoding multiple values to the same value results in corresponding variable labels being combined. Recoding to NA results in corresponding labels to be dropped since NA should not be labelled in standard haven_labelled variables.

The underlying recoding takes place using dplyr::case_match(), which takes tilde or formula notation.

Usage

lbl_recode(.x, ..., .sep = " / ")

Arguments

.x

The labelled vector to recode.

...

A recode specification in tilde (formula) syntax.

.sep

A string for how to separate labels that need to be combined.

Value

The re-coded vector.

Examples

lbl_recode(fruit_lbl, 1:2 ~ 1, 3 ~ 2, 4:5 ~ 3)
#> <labelled<double>[50]>
#>  [1] 2 3 3 2 1 1 1 3 1 1 1 1 3 3 1 3 3 3 1 1 2 3 3 3 3 1 1 2 1 2 1 2 1 1 3 2 3 3
#> [39] 2 1 1 1 3 3 1 3 2 1 3 1
#> 
#> Labels:
#>  value          label
#>      1 Apple / Banana
#>      2           Lime
#>      3 Orange / Peach
lbl_recode(fruit_lbl_chr, "d" ~ "z", "e" ~ NA)
#> <labelled<character>[50]>
#>  [1] c  NA z  c  a  b  a  z  a  a  b  a  NA z  b  NA NA z  b  b  c  z  NA NA z 
#> [26] b  b  c  b  c  a  c  b  a  z  c  z  z  c  a  b  b  z  NA b  NA c  a  NA a 
#> 
#> Labels:
#>  value  label
#>      a  Apple
#>      b Banana
#>      c   Lime
#>      z Orange
lbl_recode(exotic_veggies, 3:10 ~ 1, 11:20 ~ 2, 21:41 ~ 3,
                           2 ~ NA, NA ~ 99)
#> <labelled<double>[50]>: The exotic veggies are indexed by the prime numbers 2-41. No observations exist for indexes 7 and 41, but 8 unlabelled observations exist (for indexes 2, 8, 16 and 32), and 1 observations is NA
#>  [1] NA  3  1 NA  3  3  1  2  1  1  1  1  3  2 NA  1  1  2  1  2  2  1  1  2 NA
#> [26] NA NA  3  3  2  2  2  2  2  2  2  2  1  2  2 NA  2 NA  3  1  1 99  3  1  1
#> 
#> Labels:
#>  value                                         label
#>      1                 Celeriac / Chayote / Eggplant
#>      2           Endive / Fennel / Jicama / Kohlrabi
#>      3 Okra / Radicchio / Rutabaga / Taro / Zucchini